refactor: No Private and Auth
