--- src/components/UserInfo/index.tsx
@@ -1,3 +1,0 @@
-import UserInfo from './UserInfo';
-
-export { UserInfo as default, UserInfo };
