+++ src/hooks/index.ts
@@ -1,3 +1,2 @@
-export * from './auth'; // OR  './authFirebase';
 export * from './event';
 export * from './layout';
