+++ src/layout/index.tsx
@@ -1,4 +1,3 @@
-import PrivateLayout from './PrivateLayout';
 import PublicLayout from './PublicLayout';
 
-export { PublicLayout as default, PublicLayout, PrivateLayout };
+export { PublicLayout as default, PublicLayout };
