--- src/views/Auth/Login/LoginEmailView.tsx
@@ -1,124 +1,0 @@
-import { SyntheticEvent, useCallback, useState } from 'react';
-import { useNavigate } from 'react-router-dom';
-import { Button, Grid, TextField, Card, CardHeader, CardContent, InputAdornment } from '@mui/material';
-import { useAppStore } from '../../../store';
-import { AppButton, AppLink, AppIconButton, AppAlert, AppForm } from '../../../components';
-import { useAppForm, SHARED_CONTROL_PROPS, eventPreventDefault } from '../../../utils/form';
-
-const VALIDATE_FORM_LOGIN_EMAIL = {
-  email: {
-    presence: true,
-    email: true,
-  },
-  password: {
-    presence: true,
-    length: {
-      minimum: 8,
-      maximum: 32,
-      message: 'must be between 8 and 32 characters',
-    },
-  },
-};
-
-interface FormStateValues {
-  email: string;
-  password: string;
-}
-
-/**
- * Renders "Login with Email" view for Login flow
- * url: /auth/login/email
- * @page LoginEmail
- */
-const LoginEmailView = () => {
-  const navigate = useNavigate();
-  const [, dispatch] = useAppStore();
-  const { formState, onFieldChange, fieldGetError, fieldHasError, isFormValid } = useAppForm({
-    validationSchema: VALIDATE_FORM_LOGIN_EMAIL,
-    initialValues: { email: '', password: '' } as FormStateValues,
-  });
-  const [showPassword, setShowPassword] = useState(false);
-  const [error, setError] = useState<string>();
-  const values = formState.values as FormStateValues; // Typed alias to formState.values as the "Source of Truth"
-
-  const handleShowPasswordClick = useCallback(() => {
-    setShowPassword((oldValue) => !oldValue);
-  }, []);
-
-  const handleFormSubmit = useCallback(
-    async (event: SyntheticEvent) => {
-      event.preventDefault();
-
-      const result = true; // await api.auth.loginWithEmail(values);
-      if (!result) {
-        setError('Please check email and password');
-        return;
-      }
-
-      dispatch({ type: 'LOG_IN' });
-      navigate('/', { replace: true });
-    },
-    [dispatch, /*values,*/ navigate]
-  );
-
-  const handleCloseError = useCallback(() => setError(undefined), []);
-
-  return (
-    <AppForm onSubmit={handleFormSubmit}>
-      <Card>
-        <CardHeader title="Login with Email" />
-        <CardContent>
-          <TextField
-            required
-            label="Email"
-            name="email"
-            value={values.email}
-            error={fieldHasError('email')}
-            helperText={fieldGetError('email') || ' '}
-            onChange={onFieldChange}
-            {...SHARED_CONTROL_PROPS}
-          />
-          <TextField
-            required
-            type={showPassword ? 'text' : 'password'}
-            label="Password"
-            name="password"
-            value={values.password}
-            error={fieldHasError('password')}
-            helperText={fieldGetError('password') || ' '}
-            onChange={onFieldChange}
-            {...SHARED_CONTROL_PROPS}
-            InputProps={{
-              endAdornment: (
-                <InputAdornment position="end">
-                  <AppIconButton
-                    aria-label="toggle password visibility"
-                    icon={showPassword ? 'visibilityon' : 'visibilityoff'}
-                    title={showPassword ? 'Hide Password' : 'Show Password'}
-                    onClick={handleShowPasswordClick}
-                    onMouseDown={eventPreventDefault}
-                  />
-                </InputAdornment>
-              ),
-            }}
-          />
-          {error ? (
-            <AppAlert severity="error" onClose={handleCloseError}>
-              {error}
-            </AppAlert>
-          ) : null}
-          <Grid container justifyContent="center" alignItems="center">
-            <AppButton type="submit" disabled={!isFormValid()}>
-              Login with Email
-            </AppButton>
-            <Button variant="text" color="inherit" component={AppLink} to="/auth/recovery/password">
-              Forgot Password
-            </Button>
-          </Grid>
-        </CardContent>
-      </Card>
-    </AppForm>
-  );
-};
-
-export default LoginEmailView;
