+++ src/views/Home/index.tsx
@@ -1,0 +1,3 @@
+import HomeView from './HomeView';
+
+export { HomeView as default, HomeView };
