--- src/views/Welcome/index.tsx
@@ -1,3 +1,0 @@
-import WelcomeView from './WelcomeView';
-
-export { WelcomeView as default, WelcomeView };
